Update brand assets and adjust navbar logo
Replaced the previously used logo with the new 'Alabama.svg' branding asset in the navbar component to reflect our updated regional branding strategy. This includes modifying the alt text to correspond with the new logo name and significance while adjusting the image dimensions to maintain visual balance in the navbar layout.

The removal of 'next.svg' and 'vercel.svg' assets from the public directory is part of a clean-up process to remove unused brand assets, streamlining the project's asset management.